Move CFoulDrawConfiguration traits
