+++ include/football/CFoulDrawConfiguration.h
@@ -1,48 +1,14 @@
 #pragma once
 
 #include "IJsonable.h"
-#include "traits/default.h"
-#include "traits/json.h"
+#include "football/traits/CFoulDrawConfiguration.h"
 
 #include "football/types/CDrawConfiguration.h"
 #include "football/types/CFoulDrawConfiguration.h"
 #include "football/CLineup.h"
 #include "CEnumDistribution.h"
 
-namespace futsim
-{
-
-namespace football
-{
-class CFoulDrawConfiguration;
-}
-
-template <> struct default_traits<football::CFoulDrawConfiguration>
-{
-protected:
-	using stat = football::types::CDrawConfiguration::stat;
-public:
-	//! Default \copybrief football::CFoulDrawConfiguration::mAverageFouls
-	static inline constexpr stat AVERAGE_FOULS = stat{ 8195 } / 380;
-	//! Default \copybrief football::CFoulDrawConfiguration::mAverageYellowCards
-	static inline constexpr stat AVERAGE_YELLOW_CARDS = stat{ 1375 } / 380;
-	//! Default \copybrief football::CFoulDrawConfiguration::mAverageRedCards
-	static inline constexpr stat AVERAGE_RED_CARDS = stat{ 31 } / 380;
-};
-
-template <> struct json_traits<football::CFoulDrawConfiguration>
-{
-	//! JSON key for the class.
-	static inline constexpr std::string_view KEY = "Foul draw configuration";
-	//! JSON key for the \copybrief football::CFoulDrawConfiguration::mAverageFouls
-	static inline constexpr std::string_view AVERAGE_FOULS_KEY = "Average fouls";
-	//! JSON key for the \copybrief football::CFoulDrawConfiguration::mAverageYellowCards
-	static inline constexpr std::string_view AVERAGE_YELLOW_CARDS_KEY = "Average yellow cards";
-	//! JSON key for the \copybrief football::CFoulDrawConfiguration::mAverageRedCards
-	static inline constexpr std::string_view AVERAGE_RED_CARDS_KEY = "Average red cards";
-};
-
-namespace football
+namespace futsim::football
 {
 
 /**
@@ -131,6 +97,5 @@ std::string_view CFoulDrawConfiguration::DrawFoulCommitter( const CLineup& aLine
 	return *( players | std::ranges::views::drop( std::uniform_int_distribution<types::CLineup::names::size_type>{
 		0, static_cast<types::CLineup::names::size_type>( std::ranges::distance( players ) - 1 ) }( aGenerator ) ) ).begin();
 }
-} // football namespace
 
-} // futsim namespace
+} // futsim::football namespace
